requirements.txt
# Core Data Science
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep Learning & Transformers
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0

# Music Processing
miditok>=3.0.0
pretty_midi>=0.2.10
symusic>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
loguru>=0.7.0

# Development & Testing
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
ruff>=0.0.263
mypy>=1.0.0

# Notebooks (lightweight)
jupyter>=1.0.0
ipykernel>=6.22.0

# PDF utilities
pdfplumber>=0.9.0


